refactor(graph): move default error messages to constants file
